CreateOrder Update

+++ ETrade.Core/Application/DTOs/ProductDetailDto.cs
@@ -11,6 +11,6 @@ namespace Application.DTOs
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string CategoryName { get; set; }
-        public short UnitsInStock { get; set; }
+        public int UnitsInStock { get; set; }
     }
 }
