+++ ETrade.Core/Domain/Entities/Concrete/Product.cs
@@ -12,7 +12,7 @@ namespace Domain.Entities.Concrete
         public int CategoryId {get; set;}
         public int StoreId {get; set;}
         public string ProductName {get; set;}
-        public short UnitsInStock {get; set;}
+        public int UnitsInStock {get; set;}
         public decimal UnitPrice {get; set;}
         public int UnitsOnOrder {get; set;}
         public string Barcode {get; set;}
